requirements.txt
flask==3.0.3
flask-socketio==5.3.6
eventlet==0.33.3
werkzeug==3.0.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1